Only include id3v2 header, not Xing/LAME in slices (~fixes incorrect slice duration)
Probably not worth the effort to fix Xing/LAME header write when slicing.
With an the original file's (now incorrect) Xing/LAME header, decoders
seem to hold to the original file's duration.
Without it, decoders seem to estimate the file's duration as if it were
contant bitrate, which is not always the case and produces a gap between
the actual duration and estimated duration by decoders.
In order to avoid dealing with this issue we can:
- ensure linked files (extracted from videos server-side) are encoded w/ static bitrate
- fully decode the file and use its exact duration